+++ less/4.1.1/lib/Less/Parser.php
@@ -1734,9 +1734,29 @@ class Less_Parser {
 		if ( $e === null ) {
 			$this->save();
 			if ( $this->MatchChar( '(' ) ) {
-				if ( ( $v = $this->parseSelector() ) && $this->MatchChar( ')' ) ) {
-					$e = new Less_Tree_Paren( $v );
-					$this->forget();
+				$v = $this->parseSelector();
+				if ( $v ) {
+					// Support comma-separated selector lists inside parentheses,
+					// for pseudo-classes like :is(), :not(), :where(), :has().
+					// Backported from Less.js 4.2 (less.js#4290), elabx/less.php#1.
+					$selectors = [];
+					while ( $this->MatchChar( ',' ) ) {
+						$selectors[] = $v;
+						$selectors[] = new Less_Tree_Anonymous( ',' );
+						$v = $this->parseSelector();
+					}
+
+					if ( $v && $this->MatchChar( ')' ) ) {
+						if ( $selectors ) {
+							$selectors[] = $v;
+							$e = new Less_Tree_Paren( new Less_Tree_Expression( $selectors, true ) );
+						} else {
+							$e = new Less_Tree_Paren( $v );
+						}
+						$this->forget();
+					} else {
+						$this->restore();
+					}
 				} else {
 					$this->restore();
 				}